requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
drf-yasg==1.21.14
gunicorn==21.2.0
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
pandas==2.2.0
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
django-cors-headers==4.9.0
